+++ include/loc.h
@@ -1,16 +1,16 @@
 #ifndef LOC_H
 #define LOC_H
 
-#include <stdint.h>
+#include <stddef.h>
 
 
 struct Loc 
 {
 	const char *source;
 	char c;
-	uint32_t i;
-	uint32_t col;
-	uint32_t line;
+	size_t i;
+	size_t col;
+	size_t line;
 };
 
 void
